+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:flutter_training/screen/splash_screen.dart';
+import 'package:flutter_training/UI/screen/splash_screen.dart';
 
 void main() {
   runApp(
